Bug 1260894 - Remove broken LazyScriptCache r=jorendorff
Due to a shadowed variable in FixedSizeHash, the LazyScriptCache has not
been successfully adding items to the cache. This has been broken for a
while now and the lookup key and other constraints are no longer valid.
This patch removes the cache altogether until someone can re-qualify the
correctness and utility.

MozReview-Commit-ID: CJoB6OpNZen