Bug 1415742 - Move channel-prefs.js installation to moz.build. r=nalexander,ted
This was awkward because it doesn't want to end up under dist/bin/browser,
but DIST_SUBDIR is exported here. firefox.exe doesn't want to end up under
dist/bin/browser either, so we unset DIST_SUBDIR for this directory and
move the files that need it to a different directory.

MozReview-Commit-ID: Jr1RLUIj0HM