Bug 1518661 - Part 3: Adjust call-expression positions. r=jimb,jorendorff
When we mark call expressions as breakpoints, we want to make it as likely
as possible that the call has its own unique positon. The existing logic
means that it is more likely that the beginning of a call will align
with the start of an expression statement or other debuggable step point.
By using the property-access location, we're less likely to collide.

Thid also adds a new bytecodes that were missed in the original code that
added this position handling logic.

Differential Revision: https://phabricator.services.mozilla.com/D17661

[Waterfox note: The actual code changes of this patch were already included in
the preceding backporting patch for optional chaining. This patch just adds the
test changes from the original commit, at least insofar as they apply relatively
cleanly to Waterfox, respectively concern tests already present in our code
base.]