Bug 1596172 - Add perf-reftest-singletons to PGO training r=firefox-build-system-reviewers,mshal
This ought to overcome a regression from 1592981 as well as make this test suite less noisy about toolchain updates in the future.

While here, tidy up the PGO list: sort alphabetically, stop using a hardcoded test index for the Speedometer time extension, and stop using js-input/ for the virtual mappings since js-input/ is an actual directory in our repo.

Differential Revision: https://phabricator.services.mozilla.com/D52888

[Waterfox note: Waterfox currently only has bloom-basic.html in its repository,
so the other tests added by this patch are omitted. If any of the other perf-
reftest-singletons are ever backported as well, they should be added here, too.]