Bug 1705837 - Add new_Record / new_List utility functions to JS self-hosting. r=jandem
We do not allow `new` in self-hosting, except in these special cases. We can
avoid `new` in these remaining cases as well by using std_Object_create. The
created objects are internal to self-hosting algorithms and are never exposed
to user scripts.

Differential Revision: https://phabricator.services.mozilla.com/D112435

[Waterfox note: Part 2 of this bug (the removal of the MakeConstructible
intrinsic) is omitted for now, because our Legacy{Generator/Iterator} implemen-
tation still depends on it.]