Disable webrtc-gtests that don't build with C++17
The version of webrtc in the tree isn't completely compatible with C++17 due to
at the very least some usage of std:random_shuffle() in some unit tests.

Overriding the CXXFLAGS via moz.build doesn't work, because the global default
of using -std=c++17 is appended to the command line *after* flags coming from
the respective moz.build.

Backporting bug 1393119 might work (and is probably how C++17 was able to be
switched on by default for Firefox 73, even though at that time the webrtc
version in use *still* contained a few uses of std::random_shuffle()), but it's
not worth the effort, especially as it turns out that those tests never ran in
automation anyway (i.e. were purely for local use) and Firefox itself eventually
removed them in bug 1634675.

So we just disable the offending tests for now.