No Bug - Work around missing JSOP_GOTO functionality for optional chaining.
Original OptionalEmitter implementation uses JSOP_GOTO, but our legacy code
still uses SourceNotes for BaselineJIT compiling. The SourceNotes require
each JSOP_GOTO to be accompanied by a call to newSrcNote() with a proper note type,
but there is no such a type for the optional chaining yet.

To make the current BaselineJIT happy, we just emulate "break <label>" statement
with a fake label. The original algorithm for:

    <left> ?. <right>

can be represented as:

    if((foo=<left>) == null_or_undefined) {
        goto short_circuit;
    }
    foo<op>right
    goto finish;
    short_circuit: sh_c_preparation
    finish:

we change it to:

    fake_label: {
      if((foo=<left>) == null_or_undefined) {
          sh_c_preparation
          break fake_label;
      }
      foo<op>right
    }

(fortunately we can avoid the actual fake_label value generation :) )