Fix error in handling -c in msvcc.sh wrapper
If "-c" is encountered first, everything works fine, but if "-o" is encountered
*before* "-c", then the necessary retroactive adjustment of the output command
line switch from "-Fe" to "-Fo" doesn't happen.

This became noticeable when trying to turn on PGO, because with "-Fe", the file
extension in the output file is overriden by the assembler, so instead of the
special PGO object suffix the object file still gets the standard .obj
extension, causing the build to subsequently fail because it is looking for an
object file with the ".i_o" PGO extension.