Fixup backport of bug 1434007 to avoid debug assertion
The way bug 1434007 has been backported by simply declaring duplicate external
function definitions that point to the same internal function definition leads
to hitting the assertSelfHostedFunctionHasCanonicalName() assertion when running
in debug mode.

So instead, we just switch to the way it was done in the original patch for
bug 1434007 and create the respective aliases inside js::initStringClass().