Replace magic number usage for group slots by named constants
Somehow, the slot indices in the "Make sure that the properties are in the right
slots" debug check had become swapped around during backporting, which naturally
lead to Waterfox not even being able to start up in debug mode due to
immediately hitting an assertion failure.

This change fixes this and retrofits the named constants that were originally
introduced in bug 1480819 - Part 7 to avoid future errors of that sort.