Some new insights regarding the null checks in HostResolveImportedModule()
They were added based on https://repo.palemoon.org/MoonchildProductions/UXP/issues/1840.

I don't think the null check for the script and the context at the top is
actually necessary, because we're being called straight from Spidermonkey from
the self-hosted Module.js builtin, so something would have to go seriously
wrong for the JS context and the module function parameter to be missing.

The `!ms` check further down on the other hand is, at least for the time being,
necessary because of a real bug in the module loading code.

The componentes of that bug are as follows:

- Part of the functionality of the es-module-shims library seems to be based on
  dynamically loading scripts into blobs and then importing those.

- For as of now unknown reasons that are presumably specific to Waterfox, the
  particular version of the es-module-shims library as used by bing.com takes a
  wrong turn at Albuquerque and ends up spamming loads of module imports
  utilising blob: URIs.

- (Side note: It seems that once we'll natively support dynamic imports, the
  shims library will take a different code path and avoid getting in a seemingly
  endless loop of creating blob: URI imports.)

- From a very rough understanding, module loading hits the ScriptLoader code in
  two phases: First the modules are scheduled for fetching, then, once the
  dependencies have been resolved, the module graph is actually instantiated.

- Both during fetching and instantiating, the ScriptLoader receives the module
  specifiers as plain strings, which it then turns into nsIURIs via
  `ResolveModuleSpecifier()`.

- blob: URIs are processed by the nsHostObjectProtocolHandler. If the blob: URI
  is valid and alive, the protocol handler also adds a corresponding principal
  to the nsIURI object. If the URI has been revoked or is otherwise invalid,
  then the nsIURI ends up without a principal.

- As soon as a page is being reloaded, all of its blob: URIs are immediately
  invalidated (bug 898728). The ScriptLoader on the other hand isn't asked to
  immediately drop all pending script and module load requests, so to the extent
  that they are still possible (the ScriptLoader's document reference will be
  unset, which will block *some* operations from continuing), they will merrily
  continue executing until the ScriptLoader is eventually cycle-collected at
  some later point.

What therefore happens is this:

1. Some load requests for blob: URIs are successfully executed by the Script-
   Loader and a corresponding entry ends up in mFetchedModules. Because the
   blob: URI was valid at that time, it means that the corresponding nsIURI
   that is used as a key for the hashtable *does* have a principal.

2. The page is reloaded, which invalidates all blob: URIs created so far.

3. Because the es-module-shims library is creating a large and continuous stream
   of module load requests, it is almost certain that the reload happens right
   in the middle of some module loading operation, i.e. after some modules have
   already been fetched, but *before* they have actually been instantiated.

4. The module loading code gets around to asking the ScriptLoader to instantiate
   some previously fetched module graph.

5. When the ScriptLoader turns the specifiers into nsIURIs, because of 2. the
   nsIURI for the blob: URI end up *without* a Principal, which means that the
   nsIURI *won't* compare equal to an nsIURI with the same spec, but a valid
   principal.

6. Because the nsIURI is technically not the same nsIURI that was put into the
   mFetchedModules hashtable, this means that the attempt to retrieve the module
   from the hashtable unexpectedly fails, and without this added null check we'd
   be crashing on a release build.

As far as I can tell, I think that Firefox 60-ish is suspectible to the above
scenario, too, and I think they're mostly saved by the fact the on a real Fire-
fox, the es-module-shims library doesn't end up in the same degenerate edge case
of continuously spamming module load requests, so it's much less likely that an
untimely reload will hit just the right (wrong?) moment to invalidate a
previously fetched blob: URI.

I *can* reproduce the same crash in HostResolveImportedModule by holding down F5
in Firefox Nightlies on bing.com or the test case from
https://forum.palemoon.org/viewtopic.php?p=224083#p224083,
although interestingly only up until the 2018-05-13 Nightly. Unfortunately
nothing immediately obvious sticks out in
https://hg.mozilla.org/mozilla-central/pushloghtml?fromchange=87cc17943894&tochange=a382f8feaba4
so the whole thing remains somewhat of a mystery.

Based on the above bug description, a proper solution might be to actively
cancel all pending load requests when `Destroy()` is being called on the Script-
Loader (a bit like what was introduced quite a few Firefox-versions later in bug
1666724 and bug 1678774), but in the meantime the null check here also works as
a stopgap solution.

(Also somewhat interestingly, I'm still seeing a few crashes in
`HostResolveImportedModule()` in current Firefox versions even *with* the
patches for those two bugs, so some residual issues still seem to persist in
that code area.)