Bug 1629106 followup - Disable Ion compilation for logical assignment operators
The bytecode implementation of the logical assignment operators involves a
JSOP_GOTO, which, because our codebase predates bug 1595476, requires some de-
dicated handling code in IonControlFlow.cpp.

Until somebody understands the ControlFlowGenerator well enough to shoehorn
something in (in principle the code flow resembles an if-then-else block),
the easiest solution to avoid hitting the "not implemented" crash is disabling
Ion compilation if we encounter a logical assignment operator.