Issue #94 - Disable wantsFloatingTitlebar code on Big Sur 11.0 and higher.
With this code enabled Waterfox and other browsers abort with the following Objective-C exception: -[NSView needsOpaqueBackstop]: unrecognized selector sent to instance.

It isn't clear to me if this code is really needed anymore or not, but the Mozilla Bug https://bugzilla.mozilla.org/show_bug.cgi?id=1400057 describes incorrect rendering.

This change leaves the code active for MacOS versions prior to 11.0 when the problem started, in case the code is really necessary.